+++ src/controllers/controllers.ts
@@ -1,0 +1,9 @@
+import { login } from "./user/authController";
+import { register } from "./user/registerController";
+import { update } from "./user/updateController";
+
+export const controller = {
+  login,
+  register,
+  update,
+};
